chore: fix process ts
